clean up appinfoadapter
create array of original values in initializer and remove the
its other initializations.

implement own add, addAll and other functions to prevent constant
calls to notifyDataSetChanged and hopefully avoid losing reference
to items.

filter results changed to copying the original values object
(instead of copying a reference to it) to prevent potential bugs.