fix bugs when backupdir is oab's own data directory
add check to doBackup and restoreApk to determine if backupDir
is own datadir. if true, doBackup needs to set permissions on the
backed up files before zipping. restoreApk needs to copy the
apk to a temporary directory on the external storage because
`pm install` doesn't work when the file is in the data directory.

shares are also affected by this. proper permissions for the
shared uri should be set per transaction with FileProvider.