count the starting day when scheduling backups
take the day the backup was scheduled into account when calculating
the time to set its alarm. in this way, backups scheduled to run
every day will start on the same day they are set.
compare the scheduled hour to the current hour to prevent backups
getting scheduled in the past. this has the side effect that a schedule
set for some hour before the current hour will initially have the same
amount of time with intervals 1 and 2 (of days) until it it executes.

remove unused variable hourOfDay.