add support for special backups
as suggested by issue #42 (https://github.com/jensstein/oandbackup/issues/42)
this adds support for backing up things which are not identified
as packages by the system itself.
since the paths to these files need to be hardcoded there will be
a need for some checks to determine whether a given path is correct
on a given android version.