Change which class has appInfoList as a public field
There is no reason to have appInfoList as a public field in
the other classes when they can just access it where it originally
is created.
The reason for having it as a public field in the first place is
that it is too large to transfer between activities in an intent
extra. And when appInfoList is static it will also survive
configuration changes.