Handle configuration changes properly
Remove configChanges from AndroidManifest.xml so the activities are
actually recreated on configuration changes.
Add variable threadId in OAndBackup.java and BatchActivity.java to
track id of current running thread so it can be tracked if a thread
is running when the configuration changes happens.
Add method to show message again if the thread is still running.