Move methods to gather AppInfos to their own class
Since the need to gather package info is shared between scheduling
and the main activity, the method to do it should be in its own,
separate class which is then called by the two.
At the same time, a switch is added to control whether the packages
no longer installed are included in the list since scheduling only
does backups and therefore has no no use for them.