Include path to parent folder in the file browser
A file pointing to the parent folder is included in the list of
files generated by the file browser for navigating back a step.
To designate it clearly as the parent folder it is represented by
the string '..'. The place holder class ParentFile is used to give
FileListAdapter a way to distinguish between the normal files and
the parent file (this approach is loosely inspired by the approach
in the file manager of the cyanogenmod project).