Add option to toggle handling of symlinks
An option to choose whether to include the -L (dereference symlinks) flag
when using cp to back up. If it is turned on, errors only referring to
the /lib symlink will be ignored.
This should answer to issue #58.
(https://github.com/jensstein/oandbackup/issues/58)
The check for error on the /lock symlink of firefox is also brought back.