Handle getting a Context instead of an Activity
The receiver of the scheduled backups will most likely be
android.app.ReceiverRestrictedContext which cannot be cast to Activity
so in that case Context.startIntentSender should be used.
User interaction will not be handled in this case but since scheduled
backups are supposed to operate unattended this will not be a problem.
An error-free execution will depend on the values in the preferences
being correct.