Copy native libraries from apk
Uncompress native libraries from the apk and copy them to /data/app-lib
when restoring the apk of a system app. This is handled by
frameworks/base/core/jni/com_android_internal_content_NativeLibraryHelper.cpp
when installing normally but has to be handled manually if the apk is
just copied to /system/app.

This situation raises some problems. The value for
ApplicationInfo.nativeLibraryDir is not always valid. An example is
OpenWnnEngineEN.java of the OpenWnn input method provided with the
Android source. This class is hardcoded to expect the libraries in
either /data/data/jp.co.omronsoft.openwnn/lib or /system/lib. But while
the ApplicationInfo.nativeLibraryDir of the package is
/data/app-lib/OpenWnn (at least on my installation)
/data/data/jp.co.omronsoft/lib is a symbolic link to
/data/app-lib/jp.co.omronsoft.
Another problem is that the system may not have have processed the newly
copied apk in time for the copying of the libraries. In that case
trying to guess the library directory based on the apk name might be the
only solution but will probably fail in some cases (for example OpenWnn
where the apk may be called OpennWnn.apk which would then give
/data/app-lib/OpenWnn as the library directory).