Fix the time at which the backup will be run the next time in ScheduleService#onStartCommand()
Commit 2 of 2 fixing https://github.com/jensstein/oandbackup/issues/132 and https://github.com/jensstein/oandbackup/issues/159

Description of the bug (reffering to the code before this commit):

If a backup is scheduled in BootReceiver#onReceive() to be run after 15 minutes, it will repeat at the same time in the future instead of at the time which the user specified for the backup.

Description of the fix:

We fix the time at which the backup will be run the next time in ScheduleService#onStartCommand().