Trick aapt into believing we have native code
Since fdroidserver relies on the output of aapt dump badging to
determine if an apk contains native code, we use the fact that aapt looks
for subdirectories in a lib directory in order to populate its
"native-code" field to make it report that we have native code.
We simply build apks with dummy subdirectories with the proper names.

fdroid is tracking the issue here:
https://gitlab.com/fdroid/fdroiddata/issues/1366