Generate dummy library files in src/main/jniLibs
Instead of adding a temporary directory to the resources classpath, put
the dummy files into jniLibs where they are automatically on the
classpath. I couldn't get the resources classpath to included files
ending with .so, only files with no filename suffix were included, and
that made the apk uninstallable because of a check in
com_android_internal_content_NativeLibraryHelper.cpp which skips files
not ending in .so.