Add class to handle schedule data
At the moment the method for persisting schedules is to write to and
read from a preferences file. This is very hard to maintain so I would
like to refactor the code so that the data for the individual schedules
are represented by a class and persisted to a database.