Allow scheduled backups while device idle
On android api versions 23 and above the alarm manager needs to be set
with the setAndAllowWhileIdle method if the alarm should fire while the
device is idle. This means that I cannot use the setRepeating method any
longer and that in turn resulted in a small cleanup of the scheduling
logic. There's still some work to do though - I should probably
exchange the value for when a schedule was placed with a value for when
it last ran.