Update mockito version
To be able to mock AlarmManager on Android P I need to update mockito
(to get a newer bytebuddy version). And to get the newer bytebuddy
package to build properly I also need to update the android gradle
plugin version. And to get the newer android gradle plugin version to
work I also need to update gradle itself.